benchmarks/requirements.txt
# Benchmarking dependencies
faiss-cpu>=1.7.4
hnswlib>=0.8.0
annoy>=1.17.3
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0
tqdm>=4.65.0
scikit-learn>=1.3.0
